Test `fib` by summing previous values
Previously, `fib` was testing by comparing the generated sequence with a
list of members stored in `fib.edn`. This file was generated using `fib`
however so we were only testing that `fib`s output doesn't change.

I was planning to take a list of a few hundred thousand Fibonacci
numbers and compare the output of `fib` to this list but could only find
100 or so numbers from Wolfram Alpha and a bit of Googling around.

The test now splits the Fibonacci sequence into triples, and assets the
third member is the sum of the previous two. This is done for every
generated value from `n = 0` up to `n = 10000`.